requirements.txt
# Orders-Trendyol Excel Generator Requirements

# HTTP requests
requests>=2.31.0
cloudscraper>=1.2.71

# Date/time handling
python-dateutil>=2.8.0

# Configuration and environment
python-decouple>=3.8
python-dotenv>=1.0.0

# Excel generation
openpyxl>=3.1.0

# PDF generation and manipulation
PyPDF2>=3.0.0
reportlab>=4.0.0

# Additional utilities (optional but recommended)
colorama>=0.4.6          # For colored console output
tqdm>=4.66.0            # For progress bars

# Development and testing (optional)
pytest>=7.4.0           # For unit testing
black>=23.0.0           # Code formatting
flake8>=6.0.0           # Code linting

# Note: Python 3.8+ required for this application

